requirements.txt
alembic==0.9.6
asn1crypto==0.24.0
astroid==1.6.4
autopep8==1.3.5
Babel==2.5.3
BareNecessities==0.2.8
bcrypt==3.1.4
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
dominate==2.3.1
flake8==3.5.0
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-SQLAlchemy==2.3.2
Flask-User==1.0.1.3
Flask-WTF==0.14.2
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jwcrypto==0.4.2
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
passlib==1.7.1
pkg-resources==0.0.0
pycodestyle==2.3.1
pycparser==2.18
pydotenv==0.0.7
pyflakes==1.6.0
PyJWT==1.6.1
pylint==1.9.0
PyMySQL==0.8.1
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
python-jwt==3.1.0
pytz==2018.4
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
